backend/requirements.txt
# requirements.txt

# --- Web Scraping & HTTP Requests ---
selenium==4.21.0
requests==2.32.3
beautifulsoup4==4.12.3

# --- Database & Environment ---
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# --- NLP & Machine Learning (Using a more compatible version) ---
scikit-learn==1.4.2   # A version with better compatibility
numpy==1.26.4
pandas==2.2.2